Core/Commands, Fixed Pinfo Command,account command and probably solved more that deals with the gm level Changed GMLevel and Player level to uint8 instead of uint32, Should now show the correct GM Level

+++ src/server/game/Chat/Commands/Level2.cpp
@@ -312,9 +312,9 @@ bool ChatHandler::HandlePInfoCommand(const char* args)
         if (!result)
             return false;
 
-        Field *fields           = result->Fetch();
+        Field* fields           = result->Fetch();
         total_player_time       = fields[0].GetUInt32();
-        level                   = fields[1].GetUInt32();
+        level                   = fields[1].GetUInt8();
         money                   = fields[2].GetUInt32();
         accId                   = fields[3].GetUInt32();
         race                    = fields[4].GetUInt8();
@@ -338,7 +338,7 @@ bool ChatHandler::HandlePInfoCommand(const char* args)
     {
         Field* fields           = result->Fetch();
         username                = fields[0].GetString();
-        security                = fields[1].GetUInt32();
+        security                = fields[1].GetUInt8();
         email                   = fields[2].GetString();
         muteTime                = fields[5].GetUInt64();
 
@@ -377,7 +377,7 @@ bool ChatHandler::HandlePInfoCommand(const char* args)
     if (result2)
     {
         Field* fields = result->Fetch();
-        banTime       = fields[1].GetBool() ? 0 : fields[0].GetUInt64();
+        banTime       = int64(fields[1].GetBool() ? 0 : fields[0].GetUInt32());
         bannedby      = fields[2].GetString();
         banreason     = fields[3].GetString();
     }
